requirements.txt
langchain 
openai
huggingface_hub
python_dotenv
langchain_community